+++ packages/google_maps_flutter/google_maps_flutter_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.14.8
+
+* Updates Java compatibility version to 11.
+* Updates minimum supported SDK version to Flutter 3.24/Dart 3.5.
+
 ## 2.14.7
 
 * Adds `PlatformPatternItem` pigeon class to convert `PlatformPolyline.pattern`.